requirements.txt
openai
python-dotenv
langchain_openai
langchain-pinecone
langchain
langchain-community
chardet
flask
Flask-Cors